break out into seperate cli.py file
